fixed preference

+++ src/main/java/fr/client/cacount/services/preferencemanager/PreferenceManagerInterface.java
@@ -1,8 +1,10 @@
 package fr.client.cacount.services.preferencemanager;
 
+import java.math.BigDecimal;
+
 /**
  * Created by svirch_n on 01/03/17.
  */
 public interface PreferenceManagerInterface {
-    Double getRatio();
+    BigDecimal getRatio();
 }
